Add files via upload
Hello, when i checked on the site it didn't show my links so I uploaded them in the browser just in case. There shouldn't be a problem because I uploaded everything with GitHub program on time (there is proof of time in this file) and it said it was submitted. But I uploaded here as well just in case.
Kind greetings.